+++ Tests/AccessibilityTests/AccessibilityProject/Sources/TargetA/ProtocolIndirectlyReferencedCrossModuleByExtensionMember.swift
@@ -1,0 +1,7 @@
+public protocol ProtocolIndirectlyReferencedCrossModuleByExtensionMember {}
+public extension ProtocolIndirectlyReferencedCrossModuleByExtensionMember {
+    func somePublicFunc() {}
+}
+public class ProtocolIndirectlyReferencedCrossModuleByExtensionMemberImpl: ProtocolIndirectlyReferencedCrossModuleByExtensionMember {
+    public init() {}
+}